chore: Разбил один большой Showcase.tsx на много маленьких для каждого компонента
